Reduce combo term from default
I press combos at the same time, which should be within ~10-20ms.
As it stands, I occasionally accidentally trigger a combo when typing quickly, which is probably a good indication that the combo term is slightly too long currently.
Reducing the combo term slightly should alleviate this issue while still allowing the combos to trigger correctly.