Reduce tapping term to 100ms
There are two reasons for this.
Firstly, given the move back to default hold mode, this should help to keep held keys responsive.
Secondly, it is already sometimes the case that a modifier (particularly Shift) does not apply; reducing the hold duration should be of particular benefit in combatting this.
However, an especially short hold duration is not viable with heavy keys like the Gateron North Poles, so much less than this is undesirable.